requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
decorator==4.0.6
Django==1.9.2
django-celery==3.1.17
django-filter==1.0.4
django-ipware==1.1.3
djangorestframework==3.6.4
html5lib==0.9999999
importlib==1.0.3
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
kombu==3.0.37
Markdown==2.6.9
numpy==1.10.4
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
pylibmc==1.5.0
PyPDF2==1.25.1
pytz==2015.7
PyYAML==3.11
reportlab==3.2.0
requests==2.9.1
scipy==0.17.0
scikit-learn==0.17
selenium==2.51.1
simplegeneric==0.8.1
six==1.10.0
suds-jurko==0.6
traitlets==4.1.0
XlsxWriter==0.8.4
